JAM Implementer DAO Seeding
Background:
https://forum.polkadot.network/t/decentralized-voices-cohort-4-jam-implementers-dao/12001

We would like JAM DAO members (implementers only now, maybe service builders later) to provide addresses here in this repo to support a Telegram bot

https://github.com/permanence-dao/permanence-dao-services/tree/main/pdao-telegram-bot

so that the "1 implementer, 1 vote" can be powered by the registered addresses here.  The expectation is that this is simple Sybil mechanism to limit membership to just JAM implementers who are actually building JAM clients and expend the energy to at least register themselves here.